Update build instructions
Build target should be the rbf file, which is created from the sof
file.  The sof file is used to program the FPGA using a download
cable which most people will not have.  The rbf file is required to
program via u-Boot or Linux running on the ARM core, and is
generated from the sof file (so both sof and rbf files will be
built by running 'make rbf').

Signed-off-by: Charles Steinkuehler <charles@steinkuehler.net>