Add DE0_Nano_SoC_DB25 Project
Add a Quartus project for the DE0_Nano_SoC with a DE0-Nano_DB25
adapter board for connecting to DB25 Mesanet daughter-cards.

This initial version is a clone of the existing DE0_NANO_SOC
project with some files renamed.  Work remains to properly support
connecting the hostmot2 I/O pins to the DE0 GPIO ports (the Mesa
hostmot2 VHDL code assumes iobits maps 1:1 with the top-level
I/O pins (which is always the case on Mesa boards, since the FPGA
and I/O connectors are on the same board), but this does not hold
true for the DE0-Nano_DB25 adapter board so an additional layer of
mapping between iobits() and GPIO_0()/GPIO_1() is needed.

Signed-off-by: Charles Steinkuehler <charles@steinkuehler.net>